+++ resolwe_bio/tests/processes/test_diff_expression.py
@@ -190,7 +190,7 @@ class DiffExpProcessorTestCase(KBBioProcessTestCase):
                     },
                     "executor": {
                         "docker": {
-                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04-03042021",
+                            "image": "public.ecr.aws/genialis/resolwebio/base:ubuntu-22.04-14112023",
                         },
                     },
                 },
@@ -298,7 +298,7 @@ re-save feature_type "gene"
                     },
                     "executor": {
                         "docker": {
-                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04-03042021",
+                            "image": "public.ecr.aws/genialis/resolwebio/base:ubuntu-22.04-14112023",
                         },
                     },
                 },
@@ -434,7 +434,7 @@ re-save name name
                     },
                     "executor": {
                         "docker": {
-                            "image": "public.ecr.aws/s4q6j6e8/resolwebio/base:ubuntu-20.04-03042021",
+                            "image": "public.ecr.aws/genialis/resolwebio/base:ubuntu-22.04-14112023",
                         },
                     },
                 },